% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{smart_segment}
\alias{smart_segment}
\title{Smart segment}
\usage{
smart_segment(..., streak)
}
\arguments{
\item{...}{Column vectors to choose between}

\item{streak}{Minimum length of each run.}
}
\value{
Vector of the same length as any in \code{...} that indexes into the given vectors
}
\description{
Given a number of time-aligned vectors, choose a vector for each time such
that the sum of the values is maximized and each run is of length at least
\code{streak}.
}
