% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smart_segment_cpp}
\alias{smart_segment_cpp}
\title{C++ workhorse}
\usage{
smart_segment_cpp(values, streak)
}
\arguments{
\item{values}{Matrix whose columns are streams to switch between}

\item{streak}{Minimum length between switches}
}
\description{
Calculate the segmentation given a value matrix and minimum streak
}
